+++ src/StackExchange.Redis/ConnectionMultiplexer.Sentinel.cs
@@ -349,9 +349,8 @@ public partial class ConnectionMultiplexer
     }
 
     internal EndPoint? GetConfiguredPrimaryForService(string serviceName) =>
-        GetServerSnapshot()
-            .ToArray()
-            .Where(s => s.ServerType == ServerType.Sentinel)
+        _serverSnapshot // same as GetServerSnapshot, but without forcing span
+            .Where(static s => s.ServerType == ServerType.Sentinel)
             .AsParallel()
             .Select(s =>
             {
@@ -361,9 +360,8 @@ public partial class ConnectionMultiplexer
             .FirstOrDefault(r => r != null);
 
     internal EndPoint[]? GetReplicasForService(string serviceName) =>
-        GetServerSnapshot()
-            .ToArray()
-            .Where(s => s.ServerType == ServerType.Sentinel)
+        _serverSnapshot // same as GetServerSnapshot, but without forcing span
+            .Where(static s => s.ServerType == ServerType.Sentinel)
             .AsParallel()
             .Select(s =>
             {
@@ -425,9 +423,8 @@ public partial class ConnectionMultiplexer
 
     internal void UpdateSentinelAddressList(string serviceName)
     {
-        var firstCompleteRequest = GetServerSnapshot()
-                                    .ToArray()
-                                    .Where(s => s.ServerType == ServerType.Sentinel)
+        var firstCompleteRequest = _serverSnapshot // same as GetServerSnapshot, but without forcing span
+                                    .Where(static s => s.ServerType == ServerType.Sentinel)
                                     .AsParallel()
                                     .Select(s =>
                                     {