+++ docs/ReleaseNotes.md
@@ -8,7 +8,8 @@ Current package versions:
 
 ## Unreleased
 
-- Internal: revisit endpoint-snapshot implementation
+- Fix [#1458](https://github.com/StackExchange/StackExchange.Redis/issues/1458): Fixes a leak condition when a connection completes on the TCP phase but not the Redis handshake ([#2238 by NickCraver](https://github.com/StackExchange/StackExchange.Redis/pull/2238))
+- Internal: ServerSnapshot: Improve API and allow filtering with custom struct enumerator ([#2337 by mgravell](https://github.com/StackExchange/StackExchange.Redis/pull/2337))
 
 
 ## 2.6.86